Working on index support.
git-svn-id: https://svn.kenai.com/svn/jdbfdriver~jdbfdriver-source@19 3c4314c4-b677-ce32-ed35-be3e328b34d6

+++ JDBFDriver/src/com/idataconnect/jdbfdriver/index/NDX.java
@@ -145,7 +145,7 @@ public class NDX {
         FileChannel channel = randomAccessFile.getChannel();
         buf.position(0);
         buf.limit(BLOCK_SIZE);
-        channel.position(BLOCK_SIZE * blockNumber);
+        channel.position(BLOCK_SIZE * (long) blockNumber);
         while (buf.hasRemaining()) {
             channel.read(buf);
         }
